# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a order-service server",
    "version": "1.0.0",
    "title": "order Service"
  },
  "host": "localhost:3000",
  "basePath": "/api/v1/order",
  "paths": {
...
    "/update": {
      "patch": {
        "tags": [
          "Order API"
        ],
        "summary": "update order details",
        "description": "",
        "operationId": "updateOrder",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Order Object",
            "required": true,
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Order"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Error"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Order": {
      "type": "object",
      "properties": {
        "orderId": {
          "type": "string"
        },
        "productName": {
          "type": "string"
        }
      }
    }
  }
}